Remove JobQueueManager
It was only a thin wrapper with no value left. Move
its functionality down to the JobQueue object.
